Update the building without wallet
