+++ src/miner.cpp
@@ -25,7 +25,9 @@
 #include <util.h>
 #include <utilmoneystr.h>
 #include <validationinterface.h>
+#ifdef ENABLE_WALLET
 #include <wallet/wallet.h>
+#endif
 
 #include <algorithm>
 #include <queue>
@@ -824,6 +826,7 @@ void IncrementExtraNonce(CBlock* pblock, const CBlockIndex* pindexPrev, unsigned
     pblock->hashMerkleRoot = BlockMerkleRoot(*pblock);
 }
 
+#ifdef ENABLE_WALLET
 //////////////////////////////////////////////////////////////////////////////
 //
 // Proof of Stake miner
@@ -1022,3 +1025,4 @@ void StakeQtums(bool fStake, CWallet *pwallet, CConnman* connman, boost::thread_
         stakeThread->create_thread(boost::bind(&ThreadStakeMiner, pwallet, connman));
     }
 }
+#endif
